+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Abdullah-Koolmees 2024 voriconazole and posaconazole whole-body PBPK ([doi:10.1007/s13318-024-00916-1](https://doi.org/10.1007/s13318-024-00916-1)) -- a typical 73-kg adult, simulated as healthy, bacterially infected and ICU patient types with and without concomitant flucloxacillin.
+- Add Iwama 2024 febuxostat ([doi:10.1002/prp2.70032](https://doi.org/10.1002/prp2.70032)) -- Japanese pediatric patients with hyperuricemia including gout and adults who were healthy or had renal dysfunction.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.